% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snacks.R
\name{snacks}
\alias{snacks}
\title{Calculate my snacks}
\usage{
snacks(grapes, apples, takis)
}
\arguments{
\item{grapes}{mass of grapes (grams)}

\item{apples}{number of apples in a bucket}

\item{takis}{how many takis i eat}
}
\description{
Calculate my snacks
}
\examples{
snacks(grapes =2.5, apples = 6, takis = 1)
}
